requirements.txt
pip~=24.2
attrs~=24.2.0
PySocks~=1.7.1
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.4
pyobjc-core~=10.3.2
pyobjc-framework-Cocoa~=10.3.2
pyobjc-framework-Quartz~=10.3.2
PyMsgBox~=1.0.9
pyperclip~=1.9.0
MouseInfo~=0.1.3
PyScreeze~=1.0.1
pytweening~=1.2.0
PyGetWindow~=0.0.9
PyAutoGUI~=0.9.54
numpy~=1.26.4
PyRect~=0.2.0